+++ scripts/distribution/ubuntu-packages/build-mainnet-deb.sh
@@ -5,12 +5,12 @@ set -euxo pipefail
 # Used environment variables.
 # - UBUNTU_VERSION (mandatory)
 # - STATIC_LIBRARIES_IMAGE_TAG (defaults to 'latest' if not given)
-# - GHC_VERSION (defaults to '8.10.4' if not given)
+# - GHC_VERSION (defaults to '9.0.2' if not given)
 
 if ! docker inspect --type=image static-node-binaries > /dev/null 2> /dev/null ; then
     # build static binaries
     export STATIC_LIBRARIES_IMAGE_TAG="${STATIC_LIBRARIES_IMAGE_TAG:-latest}"
-    export GHC_VERSION="${GHC_VERSION:-8.10.4}"
+    export GHC_VERSION="${GHC_VERSION:-9.0.2}"
     export EXTRA_FEATURES="collector"
     (cd ../../../; ./scripts/static-binaries/build-static-binaries.sh)
 fi